System initialization check

+++ system-init.sh
@@ -20,25 +20,41 @@ sudo ln -s ./checks/cluster-state.sh cs.sh
 prnt "Installing etcd on $this_host_ip"
 . install-etcd.script
 #check ssh acces before we proceed any furthers
-. checks/ssh-access.sh
+if ! can_access_ip $master_ip; then
+  err "Can not access kubernetes master@$master_ip - Not proceeding with system initialization"
+  return 1
+fi
+
 if [ ! "$this_host_ip" = "$master_ip" ]; then
   prnt "Copying etcd certs to $this_host_ip"
   sudo mkdir -p /etc/kubernetes/pki/etcd/
 
-sudo -u $usr scp -q -o StrictHostKeyChecking=no -o UserKnownHostsFile=/dev/null $master_ip:/etc/kubernetes/pki/{apiserver-etcd-client.crt,apiserver-etcd-client.key} /etc/kubernetes/pki/ || ( err "Could not copy client cert and keys from $master_ip. System initialization is not complete!" && exit 1 )
+sudo -u $usr scp -q -o StrictHostKeyChecking=no -o UserKnownHostsFile=/dev/null $master_ip:/etc/kubernetes/pki/apiserver-etcd-client.crt /etc/kubernetes/pki/ || ( err "Could not copy client cert from $master_ip. System initialization is not complete!" && exit 1 )
 
-sudo -u $usr scp -q -o StrictHostKeyChecking=no -o UserKnownHostsFile=/dev/null $master_ip:/etc/kubernetes/pki/etcd/{ca.crt,ca.key} /etc/kubernetes/pki/etcd/ || ( err "Could not copy ca credenentials from $master_ip. System initialization is not complete!" && exit 1 )
-    sudo -u $usr scp -q -o StrictHostKeyChecking=no -o UserKnownHostsFile=/dev/null $master_ip:/etc/kubernetes/manifests/{etcd.yaml,kube-apiserver.yaml} $kube_vault
+sudo -u $usr scp -q -o StrictHostKeyChecking=no -o UserKnownHostsFile=/dev/null $master_ip:/etc/kubernetes/pki/apiserver-etcd-client.key /etc/kubernetes/pki/ 
+
+if [ "$?" -ne 0 ]; then
+  err "Could not copy client key from $master_ip. System initialization is not complete!"
+  return 1
+fi
+#TODO The above check need to be repeated for all
+sudo -u $usr scp -q -o StrictHostKeyChecking=no -o UserKnownHostsFile=/dev/null $master_ip:/etc/kubernetes/pki/etcd/ca.crt /etc/kubernetes/pki/etcd/ || ( err "Could not copy ca.cert from $master_ip. System initialization is not complete!" && exit 1 )
+
+sudo -u $usr scp -q -o StrictHostKeyChecking=no -o UserKnownHostsFile=/dev/null $master_ip:/etc/kubernetes/pki/etcd/ca.key /etc/kubernetes/pki/etcd/ || ( err "Could not copy ca.key from $master_ip. System initialization is not complete!" && exit 1 )
+
+sudo -u $usr scp -q -o StrictHostKeyChecking=no -o UserKnownHostsFile=/dev/null $master_ip:/etc/kubernetes/manifests/etcd.yaml  $kube_vault
+
+sudo -u $usr scp -q -o StrictHostKeyChecking=no -o UserKnownHostsFile=/dev/null $master_ip:/etc/kubernetes/manifests/kube-apiserver.yaml $kube_vault
 
     if [ -s $kube_vault/etcd.yaml ]; then
-      cat /etc/kubernetes/manifests/etcd.yaml | base64 >"$kube_vault"/etcd.yaml.encoded
+      cat $kube_vault/etcd.yaml | base64 >"$kube_vault"/etcd.yaml.encoded
       rm $kube_vault/etcd.yaml
     else
       echo -e "\e[31metcd.yaml is empty or does not exist! System initialization not complete.\e[0m"
       exit 1
     fi
     if [ -s $kube_vault/kube-apiserver.yaml ]; then
-      cat /etc/kubernetes/manifests/kube-apiserver.yaml | base64 >"$kube_vault"/kube-apiserver.yaml.encoded
+      cat $kube_vault/kube-apiserver.yaml | base64 >"$kube_vault"/kube-apiserver.yaml.encoded
       rm $kube_vault/kube-apiserver.yaml
     else
       echo -e "\e[31mkube-apiserver.yaml is empty or does not exist! System initialization not complete.\e[0m"
